+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add internal API for retrieving group members by group title [#205](https://github.com/rokwire/groups-building-block/issues/205)
 
 ## [1.5.57] - 2022-07-26
 ### Changed